backend/pyproject.toml
[project]
name = "anime-tracker-backend"
version = "0.1.0"
description = "FastAPI backend for the Anime Tracker project."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14.0",
  "fastapi>=0.115.0",
  "fastapi-cache2>=0.2.2",
  "httpx>=0.28.0",
  "pydantic-settings>=2.6.0",
  "python-dotenv>=1.0.1",
  "sqlalchemy>=2.0.36",
  "uvicorn[standard]>=0.32.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "ruff>=0.7.0"
]

[build-system]
requires = ["setuptools>=68.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
